Program design changed (indications of the professor)
The Prolog program design was changed based on indications given by the
professor of the practical classes on 2016/12/15. He advised to use a
matrix to the number of votes. The 'SeriesDataGenerator' has been
adjusted to the new design and had some bug fixes and improvements. The
majority of the changes have been done on the practical class of
2016/12/15.