PA2 complete
Full redesign of "TV Programming Optimizer"/"TV Programation Optimizer"
and some data files and the report added.  The predicate
"contiguousSeries()" is not doing anything visible, but is almost
correct (see the code and the report). Most of the assignment was done
on 2016/12/21 and 2016/12/22 with some enhancements on 2016/12/23. In
the earlier days, what was done wasn't working properly.